requirements.txt
# Earth Digital Immune System (EDIS) - Backend Requirements

# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
python-dotenv==1.0.0
reportlab==4.0.7

# AI/ML Dependencies
groq==0.4.1
openai==1.3.7

# Database (if needed)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
